requirements.txt
flask
google.generativeai
pdf2jpg
gunicorn